% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCCCTimeSeries.R
\name{extractCCCTimeSeries}
\alias{extractCCCTimeSeries}
\title{Extract chromatic coefficients and their statistics for an array of JPEG files}
\usage{
extractCCCTimeSeries(rmskList, mIndex, paths)
}
\arguments{
\item{rmskList}{rasters of mask as list}

\item{mIndex}{A vector of integer numbers as the index of mask files, same length as paths. This vector shows which mask should be used with which JPEG file.}

\item{paths}{paths to the JPEG files}
}
\value{
This function returns statistical metrics for each color channel. The function returns NULL, if dimensions do not agree.
}
\description{
This function apply a list of mask matrices to a vector of jpeg images and extract statstical metrics for each chromatic coefficients on R, G and B.
}
\keyword{bcc}
\keyword{chromatic}
\keyword{coefficients}
\keyword{exract}
\keyword{gcc}
\keyword{rcc}
