% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCLArray.R
\name{getCLArray}
\alias{getCLArray}
\title{Center line arrary of an array of image}
\usage{
getCLArray(files)
}
\arguments{
\item{files}{path to the JPEF files}
}
\description{
This function returns CLI array for vector of JPEG files
}
\examples{

f <- system.file(package = 'xROI', 'dukehw.jpg')
cli <- getCL(f)

}
\keyword{CLI}
\keyword{Center-line}
\keyword{image}
