% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUI.R
\name{getUI}
\alias{getUI}
\title{UI function}
\usage{
getUI()
}
\value{
the user intrface module of the shiny app.
}
\description{
This funciton constructs the user interface of the shiny app.
}
\keyword{App}
\keyword{Shiny}
\keyword{UI}
