% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Launch.R
\name{Launch}
\alias{Launch}
\title{Launch xROI app}
\usage{
Launch(inputDir = NULL, Interactive = FALSE)
}
\arguments{
\item{inputDir}{a character string.  path to the input directory.}

\item{Interactive}{logical variable to force an interactive session}
}
\value{
this should be run in an interactive R session
}
\description{
This function launches the app by opening the default web browser.
}
\examples{

#Launch xROI app
xROI::Launch()


}
\keyword{App}
\keyword{Launch}
\keyword{Run}
\keyword{xROI}
