% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhenocamFilenames.R
\name{parsePhenocamFilenames}
\alias{parsePhenocamFilenames}
\title{Parse Phenocam filenames}
\usage{
parsePhenocamFilenames(filepaths)
}
\arguments{
\item{filepaths}{a character vector of filenames}
}
\value{
a datatable containing filenames, with site name, date and timing
}
\description{
This function parse filename to extract sitename, date and timing of the images based on the phenocam naming convention.
}
\keyword{Filename}
\keyword{Parse}
