% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotJPEG.R
\name{plotJPEG}
\alias{plotJPEG}
\title{Plot JPEG image}
\usage{
plotJPEG(path, add = FALSE, xlim = NULL, ylim = NULL)
}
\arguments{
\item{path}{a character string. path to the JPEG file to be plotted.}

\item{add}{logical. logical variable whether to add the image to the existing graphics.}

\item{xlim}{numeric vector of lenght 2, x axis range}

\item{ylim}{numeric vector of lenght 2, y axis range}
}
\value{
This function returns statistical metrics for each color channel. The function returns NULL, if dimensions do not agree.
}
\description{
This funciton plots a jpeg image as a raster given image path.
}
\examples{
f <- system.file(package = 'xROI', 'dukehw.jpg')
if (interactive() ) plotJPEG(f)

}
\keyword{bcc}
\keyword{chromatic}
\keyword{coefficients}
\keyword{exract}
\keyword{gcc}
\keyword{rcc}
