% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROIx.R
\name{runROIx}
\alias{runROIx}
\title{Launch ROIx app}
\usage{
runROIx(inputDir = ".")
}
\arguments{
\item{inputDir}{path to the input directory as a character string.}
}
\description{
This function launch the app by opening the default web browser.
}
\examples{

#Launch ROIx app
runROIx()

}
\keyword{App}
\keyword{Launch}
\keyword{ROIx}
\keyword{Run}
