% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPolygon.R
\name{drawPolygon}
\alias{drawPolygon}
\title{Interactive drawing of a polygon}
\usage{
drawPolygon(col = "#80303080", lty = 1, ...)
}
\arguments{
\item{col}{color value of the polygon polygon}

\item{lty}{lty variable as line type}

\item{...}{passing graphical parameters tp}
}
\description{
This function provides an interactive tool for drawing of polygons by user clicks on the plot
}
\examples{

#user can click to add vertices, pressing the Escape key would end it.
if(interactive()){
   drawPolygon()
}

}
\keyword{interactive}
\keyword{plot}
\keyword{polygon}
