% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Launch.R
\name{Launch}
\alias{Launch}
\title{Launch xROI app}
\usage{
Launch(inputDir = NULL)
}
\arguments{
\item{inputDir}{path to the input directory as a character string.}
}
\description{
This function launch the app by opening the default web browser.
}
\examples{

#Launch xROI app
xROI::Launch()


}
\keyword{App}
\keyword{Launch}
\keyword{Run}
\keyword{xROI}
