% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseROI.R
\name{parseROI}
\alias{parseROI}
\title{Parse ROI list file}
\usage{
parseROI(roifilepath)
}
\arguments{
\item{roifilepath}{a character string. path to the ROI file}
}
\value{
a list. ROI list file as a list.
}
\description{
This function reads the ROI list file and returns it as a list variable
}
\examples{

f <- system.file(package = 'xROI', 'example/ROI/example_DB_0001_roi.csv')
roi <- parseROI(f)

}
\keyword{ROI}
\keyword{ROIList}
