% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCLArray.R
\name{plotCLArray}
\alias{plotCLArray}
\title{Plot CLI array}
\usage{
plotCLArray(clArray, bands = 1:3)
}
\arguments{
\item{clArray}{a numeric array. A 3D array of CLI (HxWx3)}

\item{bands}{an integer vector. integer vector of length 3, showing bands to be plotted}
}
\value{
invisibly returns the dimension of the plotted image
}
\description{
This function plots a CLI array on the graphics.
}
\examples{
f <- system.file(package = 'xROI', 'dukehw-cli.jpg')
jp <- jpeg::readJPEG(f)
if(interactive())plotCLArray(jp)

}
\keyword{bcc}
\keyword{chromatic}
\keyword{coefficients}
\keyword{exract}
\keyword{gcc}
\keyword{rcc}
